requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
coverage==7.8.0; python_version >= '3.9'
django==5.2; python_version >= '3.10'
django-debug-toolbar==5.1.0; python_version >= '3.9'
django-rest-framework==0.1.0
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
factory-boy==3.3.3; python_version >= '3.8'
faker==37.1.0; python_version >= '3.9'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
pyjwt==2.9.0; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.24.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
whitenoise==6.9.0; python_version >= '3.9'
